requirements.txt
datasets
faiss-cpu
fastapi
jsonargparse[signatures]
lightning
regex
tensorboard
torch
tqdm
transformers[ja]
ujson
